Reject wrong called AE Title if configured.
DcmSCP has two different ways to handle the Called AE Title used by
the connecting remote AE: It can be configured to respond with the
Called AE Title (default behaviour) assuming that this would be DcmSCP's
own AE Title, or the DcmSCP user can set an AE Title that the remote AE
must use and otherwise reject the association.

This rejection had not been implemented, i.e. DcmSCP always responded
with the Called AE Title no matter what the user configured. This
rejection is now in place.

The dcmrecv tool being based on DcmSCP also suffered from this bug. The
default behaviour was to set an explicit AE Title but dcmrecv did then
not enforce it. dcmrecv has been updated to keep the old behavior, i.e.
to respond with the Called AE Title the remote AE used in order not to
to not break existing implementations that update to the new version.
However, if now the pre-existing option --aetitle is explicitly used,
the related AE Title will be enforced.